Make miltiple fitness line more lightweight (fix cycles in import and find_best_running_fitness)
